Staking/website backend/requirements.txt
appdirs==1.4.4
asgiref==3.7.2
bitarray==2.6.0
bitstring==4.0.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.10.0
djangorestframework==3.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
pycparser==2.21
PyNaCl==1.5.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tonsdk==1.0.13
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
websocket-client==1.6.2
